Player can harvest all entities on a tile. Entity is killed and give all

+++ Assets/Scripts/Entities/EntitySO.cs
@@ -6,7 +6,7 @@ using System;
 using RotaryHeart.Lib.SerializableDictionary;
 
 [Serializable]
-public class ResourcesToHarvest : SerializableDictionaryBase<ResourceType, float>
+public class ResourcesToHarvest : SerializableDictionaryBase<ResourceType, int>
 {
 }
